sync `LegacyPIC` pins with imported `BhyveAtPic`
Userspace PIC IRQ pins are tracked by the `LegacyPIC` type. This type
counts the number of times a pin is asserted, so that we can only
transition the virtual pin's state on the rising/falling edge. However,
pin level counts are _also_ tracked in kernelspace by Bhyve, and we
treat these levels as authoritative, rather than the `LegacyPIC`s
userspace view of the counts. This means there's a potential race on
imports, if `BhyveAtPic` is imported before we import a deevice such as
a UART with a pending IRQ. If we import the `BhyveAtPic` first, then
kernelspace will see the IRQ pin that was asserted on the source VM, but
the UART device will try and assert its IRQ using `LegacyPIC`, which
won't know that we already imported that pin as asserted.

This commit changes this so that importing `BhyveAtPic` also sets the
userspace pin counts on `LegacyPic` to line up with the kernel's view of
the world.