smf: set `startd/critical_failure_count` to 0
Currently, the SMF manifest for Propolis allows it to be restarted by
`startd` on crashes. This is not great, because if a `propolis-server`
process panics, the VMM state is (currently) non-recoverable, and the
restarted Propolis will just believe it hasn't got a VM instance at all.

This commit changes the `startd/critical_failure_count` property in the
manifest to 0, which should result in the service being placed in
maintainance if it crashes a single time.